requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
dbt-postgres==1.8.2
psycopg2-binary==2.9.9
fastapi==0.115.2
uvicorn==0.32.0